requirements.txt
# Selenium WebDriver for browser automation
selenium>=4.0.0
gherkin-official>=38.0.0